requirements.txt
aiosmtplib==1.1.7
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.3
Faker==19.11.1
fontawesomefree==6.4.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.4
messages==0.8.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
validus==0.3.0
